requirements.txt
# ─── Core ───────────────────────────────────────────────────
opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.1.0         # YOLOv8 person detection

# ─── Face Recognition ───────────────────────────────────────
face-recognition>=1.3.0    # 128-D face embeddings (dlib)
dlib>=19.24.0

# ─── Queue ──────────────────────────────────────────────────
redis>=5.0.0               # Redis client (queue between camera & workers)

# ─── Database ───────────────────────────────────────────────
psycopg2-binary>=2.9.9     # PostgreSQL driver
                           # (use psycopg2 instead of psycopg2-binary in production)

# ─── Config ─────────────────────────────────────────────────
python-dotenv>=1.0.0       # Load .env file

# ─── Dashboard (optional, Phase 4) ──────────────────────────
streamlit>=1.30.0
pandas>=2.0.0

# ─── Optional ───────────────────────────────────────────────
# onnxruntime>=1.16.0      # if using insightface ArcFace
# insightface>=0.7.3       # alternative face recognition backend
# httpx>=0.25.0            # Telegram notifications
# torch>=2.0.0             # if using GPU-backed models
# torchvision>=0.15.0
